requirements.txt
absl-py==1.4.0
altair==5.2.0
altgraph==0.17.3
anyio==3.6.2
asgiref==3.5.2
asttokens==2.2.1
astunparse==1.6.3
attrs==23.2.0
backcall==0.2.0
blinker==1.7.0
CacheControl==0.13.1
cachetools==5.3.1
celluloid==0.2.0
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
cmake==3.28.0
colorama==0.4.6
contourpy==1.0.7
copier==8.1.0
cryptography==41.0.4
cvzone==1.6.1
cycler==0.11.0
decorator==5.1.1
Django==4.1
dm-tree==0.1.8
dunamai==1.18.0
et-xmlfile==1.1.0
executing==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
firebase-admin==6.2.0
Flask==3.0.2
flatbuffers==23.5.26
flet==0.9.0
flet-core==0.9.0
flet-runtime==0.9.0
fonttools==4.38.0
funcy==2.0
future==1.0.0
gast==0.4.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.12.0
google-api-python-client==2.101.0
google-auth==2.19.1
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
h5py==3.8.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.24.1
idna==3.4
importlib-metadata==6.11.0
ipython==8.14.0
itsdangerous==2.1.2
jax==0.4.12
jedi==0.18.2
Jinja2==3.1.2
jinja2-ansible-filters==1.3.2
joblib==1.2.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jws==0.1.3
keras==2.15.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.0.7
mysql-connector-python==8.0.29
namex==0.0.7
numpy==1.23.5
oauth2client==3.0.0
oauthlib==3.2.2
opencv-python==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pathspec==0.11.2
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.4.0
plotly==5.21.0
plumbum==1.8.2
prompt-toolkit==3.0.36
proto-plus==1.22.3
protobuf==4.23.2
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.12
pydeck==0.8.1b0
pygame==2.1.2
Pygments==2.15.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
PyJWT==2.8.0
pyparsing==3.0.9
pypng==0.20220715.0
pyscroll==2.29
python-dateutil==2.8.2
python-jwt==2.0.1
PyTMX==3.31
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyyaml-include==1.3.1
qrcode==7.4.2
questionary==2.0.0
ratelim==0.1.6
referencing==0.32.1
repath==0.9.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.7.0
rfc3986==1.5.0
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
smmap==5.0.1
sniffio==1.3.0
sqlparse==0.4.2
stack-data==0.6.2
streamlit==1.29.0
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.0
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2022.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.16
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.6
websocket-client==1.5.2
websockets==11.0.3
Werkzeug==3.0.2
wrapt==1.14.1
zipp==3.17.0
